requirements.txt
# Add required packages here
torch>=1.7.1
torchvision>=0.8.2
ftfy
regex
tqdm
numpy>=1.19.0
Pillow>=8.0.0
faiss-cpu>=1.7.2
clip @ git+https://github.com/openai/CLIP.git
pandas
matplotlib
joblib